requirements.txt
redis
huey
PyYAML
docker
